photoedit_mvp/requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
pillow>=9.0.0
streamlit>=1.16.0
torch>=1.10.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
